Call SILE.nodefactory instantiators as functions
The sloppy mismatch between objects that could be referenced directly
and function instantiators went away in SILE v0.10.0 in favor of always
using instantiation functions (and a new object model to go with them).

There should have been a deprecation for this that threw a warning
first, but evidently that didn't work.

This commit is the minimum required to to the same thing, but there is
probably a better way to achieve the desired effect.